Activity 1 step 3 completed and step 2 new changes

+++ RoboticLeg/ex03/valid.m
@@ -1,6 +1,6 @@
 v2 = JBF*dq
 
-if (v == v2)
+if (v == double(v2))
     disp('CORRECT');
 else
     disp('INCORRECT');
